chore: Encrypt site build params #4464
